Remove Malloc

+++ crc.h
@@ -3,5 +3,5 @@ typedef struct crc
     unsigned char digest[2];
 }Crc;
 
-Crc generateCRC(unsigned char *message, int size);
-Crc validateCRC(unsigned char *message, int size);
+Crc generateCRC(unsigned char *message, size_t size);
+Crc validateCRC(unsigned char *message, size_t size);
